exceptions: Moved the module containing exceptions for the resolver module to a new utils `exceptions.utils` package
