+++ src/blasmodcli/exceptions/utils/__init__.py
@@ -1,0 +1,1 @@
+from .resolver import DependencyResolutionException, UnresolvableDependency
